requirements.txt
pytest>=8.0.0
python-dotenv>=1.0.0
telethon>=1.38.0
pybit>=5.9.0
streamlit>=1.44.0
flask>=3.1.0
